# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: h26s_03 API
  version: 0.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/tasks:
    post:
      operationId: createTask
      summary: タスクを作成する
      description: 指定したメンバーを担当者として、新しいタスクを作成します。
      requestBody:
        required: true
        description: 作成するタスクの内容です。
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: タスクの作成に成功しました。
components:
  schemas:
...
    CreateTaskRequest:
      type: object
      description: タスク作成時に送る内容です。
      required:
        - title
        - member_id
      properties:
        title:
          type: string
          description: 作成するタスク名です。
        member_id:
          type: integer
          format: int64
          description: 担当者にするメンバーIDです。